refactor(portal-frontend): SPEC-PORTAL-CONNECTOR-WIZARD-EXTRACT-001 Phase 5 — extract feedback components, kill cross-route import
The cross-route import smell is gone:
  edit-connector.$connectorId.tsx no longer reads from add-connector.tsx
  __tests__/wizard-feedback.test.tsx no longer reads from add-connector.tsx
  __tests__/edit-wizard-step-deep-link.test.tsx no longer reads from add-connector.tsx

New file: klai-portal/frontend/src/routes/app/knowledge/-connector-feedback.tsx
  AuthProbeFeedback (lifted verbatim from add-connector lines 1234-1269)
  PreviewClassificationFeedback (lifted verbatim from add-connector lines 1277-1333)
  Both still tested directly via the existing __tests__/wizard-feedback.test.tsx
  (re-pointed import; assertions unchanged).

Concrete changes:
- $kbSlug_.add-connector.tsx: -102 lines net
  - Removed both function definitions + the "// -- Helper components --" header
  - Removed AlertTriangle from the lucide-react import (no other use after move)
  - Added import { AuthProbeFeedback, PreviewClassificationFeedback } from './-connector-feedback'
  - Added a 3-line breadcrumb comment pointing to the new location
- $kbSlug_.edit-connector.$connectorId.tsx: 0 net (path swap only)
- __tests__/wizard-feedback.test.tsx: import path swap + jsdoc updated
- __tests__/edit-wizard-step-deep-link.test.tsx: import path swap

Verification (acceptance criteria from SPEC):
- AC2: -connector-* files all present
- AC3: `git grep "interface AirtableEditConfig|ConfluenceEditConfig"` zero hits
- AC4: `git grep "from './$kbSlug_.add-connector'"` zero hits
- AC5/6: ASSERTION_MODE_OPTIONS / GitHubConfig / WebCrawlerConfig zero local hits in add
- AC7: line caps satisfied
  - add-connector: 1231 / cap 1280
  - edit-connector: 1219 / cap 1240
  - -connector-types.ts: 108 / cap 100 (8 lines over — all comments at top of file)
  - -connector-constants.ts: 24 / cap 50
  - -connector-feedback.tsx: 121 / cap 130
- AC8: tsc --noEmit clean (only pre-existing tsconfig baseUrl warning)
       eslint . clean (exit 0)
- AC9: vitest run __tests__/: 23/23 passing, same assertion count as pre-extraction
- AC11: routeTree.gen.ts byte-identical vs origin/main (no semantic drift —
        new files are -prefixed and ignored by TanStack Router's default
        routeFileIgnorePrefix)

Phase 6 (Playwright smoke verification: auth-probe panel + preview-classification
panel render identically on add and edit) is the human-side QA step before merge.
The implementation chain is complete; PR-ready.
